テスト用モッククラスをテストターゲットに移動
